fix vscodes notice and rename base -> processbase
